Synopsis
About the project:
The purpose of this project is to build a basic model system of water irrigation
and integrate this into an advanced model which can be produced and successfully
be implemented into a real world. Today the world face the problems of water
scarcity as the world suffers from lack of rain and lack of groundwater throughout
the year thus requiring us to change our old methods into modern green sustainable
ways.
Reason to choose the topic:
Agriculture is developing from mechanized by simpler methods in the twentieth
century to being automated in the second decade of the 21 st century. Today the main
requisite of a successful agriculture model is its production value, its accuracy,its effect
on the environment and most importantly its production cost. To reach these
prerequisites, automated systems must be introduced and a framework of mechanics
and technology must be achieved.
In India, agriculture sector is the largest sector to its economy, but when it comes to
technology we find that it is still very deficient. This is because of the high cost it takes
than the presently used conventional methods. Now a day’s there is huge
enhancement in technology which plays a vital impact on various fields like healthcare,
education, etc. The problem with Agriculture is that most farmers are unable to afford
these technology from big brands due to their large price marks. This result into a
devastating fact that most of the irrigation systems are operated manually. This is a
risky method as it may give rise to problems of wasting water, over watering, lack of
watering. The available techniques that are used for irrigation are drip irrigation and
sprinkler irrigation etc. These techniques are to be combined with IoT so that we can
eliminate human error and analyze the soil through various sensors and conclude if
there is a requirement for watering.
Objective and scope of the project:
In this project, we have attempted to build a smart irrigation systems using IoT
and achieve the desired objectives of this system. The project is facilitated to simplify
the irrigation systems by installing the system, and increase crop performance by
reducing overwatering from saturated soil. The idea is to eliminate human error by
assigning an engine to realize the need of watering and thus preserving water from
wasting. In recent years, with a decrease in the cost of internet, there is a hope of light
to provide the farmers with a fairly low-cost but an effective system to help them
irrigate their fields as per requirements through automatic engine and thus decreasing
cost of watering and manpower. By solving these problems we might also be able to
change this cycle of wasting water and help prevent water scarcity.