Add files via upload
About the project:
The purpose of this project is to build a basic model system of water irrigation
and integrate this into an advanced model which can be produced and successfully
be implemented into a real world. Today the world face the problems of water
scarcity as the world suffers from lack of rain and lack of groundwater throughout
the year thus requiring us to change our old methods into modern green sustainable
ways.
Reason to choose the topic:
Agriculture is developing from mechanized by simpler methods in the twentieth
century to being automated in the second decade of the 21 st century. Today the main
requisite of a successful agriculture model is its production value, its accuracy,its effect
on the environment and most importantly its production cost. To reach these
prerequisites, automated systems must be introduced and a framework of mechanics
and technology must be achieved.
In India, agriculture sector is the largest sector to its economy, but when it comes to
technology we find that it is still very deficient. This is because of the high cost it takes
than the presently used conventional methods. Now a day’s there is huge
enhancement in technology which plays a vital impact on various fields like healthcare,
education, etc. The problem with Agriculture is that most farmers are unable to afford
these technology from big brands due to their large price marks. This result into a
devastating fact that most of the irrigation systems are operated manually. This is a
risky method as it may give rise to problems of wasting water, over watering, lack of
watering. The available techniques that are used for irrigation are drip irrigation and
sprinkler irrigation etc. These techniques are to be combined with IoT so that we can
eliminate human error and analyze the soil through various sensors and conclude if
there is a requirement for watering.
Objective and scope of the project:
In this project, we have attempted to build a smart irrigation systems using IoT
and achieve the desired objectives of this system. The project is facilitated to simplify
the irrigation systems by installing the system, and increase crop performance by
reducing overwatering from saturated soil. The idea is to eliminate human error by
assigning an engine to realize the need of watering and thus preserving water from
wasting. In recent years, with a decrease in the cost of internet, there is a hope of light
to provide the farmers with a fairly low-cost but an effective system to help them
irrigate their fields as per requirements through automatic engine and thus decreasing
cost of watering and manpower. By solving these problems we might also be able to
change this cycle of wasting water and help prevent water scarcity.

Methodology:
The proposed system consists of three sensors – rain, soil-moisture, and
temperature, humidity sensors. All the sensors along with the pump and the relay
are connected with the microcontroller. After the sensor read the data from the
environment, the microcontroller decides if the pump needs to be set ON/OFF. The
data collected are also trained and tested using AR and ARIMA model, which give
us an idea of how much the experimental value differ from the predicted values.

System Requirement:
Functional requirements:
Our system works through different stages and this stage can be found in:
1. If the moisture sensor is dry in the line (its locations), the system will be checked
by the rain sensor if there is rain, the system will not work because no need to
irrigate at the same time of rain, otherwise the system will check the temperature
sensor with light sensor if the temperature is high and the percentage of light Is high
as well then the system will not work because it is not the right time for irrigation
process because the water will easily evaporate.
2. If the temperature is low and the light is low and there is no rain but the moisture
sensor is dry then signal will be sent to the controller to open the valve and pump.
3. If water level is low in the tank then the system will shut down automatically and
send SMS to the user, by using water level sensor.
4. When the system is ON by using flow meter sensor connected to LCD we can know
the amount of water goes from tank to each line so if there is a leak in the pipe we can
know from LCD.
Non—Functional Requirements:
The materials required for this project must be easy to install to implement a
successful project. In addition, materials should be easy to connect with each other to
build this project and become more effective. Also, the materials of this project must be
easy to replace it in case of any damage.

FUTURE ENHANCEMENT:
The following are some features we would like to add to this project:
1. A Mobile Application – With increase in usage of this device we would like our users
to be equipped with a mobile application to monitor and control this system.
2. Image Processing – Use of this technology will allow us to get a new aspect to the
information that we are collecting and we would be able to know more about the
condition of the crop.
3. Machine Learning – Using these algorithms can help us predict our data beforehand
and thus making us able to schedule and plan for our next steps to irrigation or
agriculture.
CONTRIBUTION OF THE PROJECT:

● It reduces water consumption to a greater extent.
● It needs minimal maintenance .The power consumption has been reduced very
much.
● The system can be used in greenhouses. The System is very useful in areas where
water scarcity is a major problem .
● The crop productivity increases and the wastage of crops is very much reduced
using this irrigation system.
● The developed system is more helpful and gives more feasible results.
ACKNOWLEDGMENT:
We would like to take this opportunity to express our hearty gratitude and sincere
thanks towards our guide Dr. Neeraj Jain for his invaluable assistance in our
project.